Add files via upload
This one has a booking window added into the interface, which only shows up when you click on one of the available times. If you click an already booked time, it ignores the event. This gives us the info for filling out the json schedule document, which should be stored on the captura.com/booking page, or some other page name.